pyproject.toml
[tool.poetry]
name = "django-task-manager"
version = "0.1.0"
description = "Task manager with telegram integration"
authors = ["AlanZhuman <alan.zhuman.05@gmail.com>"]
license = "None"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
celery = "^5.4.0"
django-rest-framework = "^0.1.0"
django-celery-beat = "^2.7.0"
psycopg2-binary = "^2.9.10"
django-extensions = "^3.2.3"
drf-spectacular = {extras = ["sidecar"], version = "^0.27.2"}
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
